♻️ (못먹는음식):가이드에 맞게 단일책임원칙 준수
