+++ apps/web/src/components/Allergy/__tests__/AllergyCheck.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 
-import AllergyCheck from '../AllergyCheck';
+import AllergyCheckContainer from '../AllergyCheckContainer';
 
 describe('AllergyCheck 컴포넌트 테스트', () => {
   const mockOnSubmit = jest.fn();
@@ -10,14 +10,14 @@ describe('AllergyCheck 컴포넌트 테스트', () => {
   });
 
   it('모든 카테고리 섹션 렌더링', () => {
-    render(<AllergyCheck onSubmit={mockOnSubmit} />);
+    render(<AllergyCheckContainer onSubmit={mockOnSubmit} />);
 
     expect(screen.getByText('해산물류')).toBeInTheDocument();
     expect(screen.getByText('동물성 식품')).toBeInTheDocument();
   });
 
   it('각 카테고리의 모든 항목 렌더링', () => {
-    render(<AllergyCheck onSubmit={mockOnSubmit} />);
+    render(<AllergyCheckContainer onSubmit={mockOnSubmit} />);
 
     // 해산물류 항목들
     expect(screen.getByText('어류')).toBeInTheDocument();
@@ -35,13 +35,13 @@ describe('AllergyCheck 컴포넌트 테스트', () => {
   });
 
   it('제출 버튼 렌더링', () => {
-    render(<AllergyCheck onSubmit={mockOnSubmit} />);
+    render(<AllergyCheckContainer onSubmit={mockOnSubmit} />);
 
     expect(screen.getByRole('button', { name: /submit/i })).toBeInTheDocument();
   });
 
   it('선택된 항목이 없을 때 onSubmit 호출', async () => {
-    render(<AllergyCheck onSubmit={mockOnSubmit} />);
+    render(<AllergyCheckContainer onSubmit={mockOnSubmit} />);
 
     const submitButton = screen.getByRole('button', { name: /submit/i });
     fireEvent.click(submitButton);
@@ -52,7 +52,7 @@ describe('AllergyCheck 컴포넌트 테스트', () => {
   });
 
   it('선택된 항목이 있을 때 onSubmit 호출', async () => {
-    render(<AllergyCheck onSubmit={mockOnSubmit} />);
+    render(<AllergyCheckContainer onSubmit={mockOnSubmit} />);
 
     // 어류 선택
     const fishButton = screen.getByText('어류');
@@ -72,7 +72,7 @@ describe('AllergyCheck 컴포넌트 테스트', () => {
   });
 
   it('항목 선택 및 선택 해제 처리', async () => {
-    render(<AllergyCheck onSubmit={mockOnSubmit} />);
+    render(<AllergyCheckContainer onSubmit={mockOnSubmit} />);
 
     // 어류 선택
     const fishButton = screen.getByText('어류');
@@ -91,7 +91,7 @@ describe('AllergyCheck 컴포넌트 테스트', () => {
   });
 
   it('여러 선택 상태 유지', async () => {
-    render(<AllergyCheck onSubmit={mockOnSubmit} />);
+    render(<AllergyCheckContainer onSubmit={mockOnSubmit} />);
 
     // 여러 항목 선택
     const fishButton = screen.getByText('어류');
