+++ apps/web/src/components/Allergy/index.ts
@@ -1,1 +1,3 @@
-export { default as AllergyCheck } from './AllergyCheck';
+export { default as AllergyCheckContainer } from './AllergyCheckContainer';
+export { default as AllergyCheckPresenter } from './AllergyCheckPresenter';
+export { default as useAllergyCheck } from './useAllergyCheck';
